requirements.txt
crawl4ai
langchain
langchain-community
langchain-ollama
chromadb
playwright
beautifulsoup4
python-dotenv
sentence-transformers
boto3
openai
httpx
langchain-openai
fastapi
uvicorn
